Completely refactor the src directory and code structure to follow SOLID
The changes are an attempt to follow the SOLID principle, by splitting
the modules in multiple fewer modules, each with their own functionality.
Modules that are helping the main modules by providing extra
functionality are moved into their separate folder, and mediator modules
are created to reduce tight coupling. Moreover, the open closed
principle is now adopted, making it possible to remove an entire
page without affecting the rest of the application, while also
allowing the possibility to create and add new pages without having
to modify the already existing code.